Work version of program, with error struct

+++ src/parallel_executor/ParallelClientExecutor.h
@@ -12,7 +12,7 @@
 
 class ParallelClientExecutor: public InterfaceDeviceClient{
 public:
-    void ParallWorker(size_t, size_t);
+    void ParallWorker(int, int);
 };
